i-pac-main/requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.2
colorama==0.4.1
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.6
Django==2.1.7
h5py==2.9.0
imutils==0.5.2
kiwisolver==1.0.1
llvmlite==0.28.0
matplotlib==3.0.3
numba==0.43.1
numpy==1.16.2
opencv-contrib-python==4.0.1.24
opencv-python==4.0.0.21
paramiko==2.4.2
path.py==10.6
Pillow==5.4.1
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
scipy==1.2.1
seed==0.11.3
six==1.12.0
tqdm==4.31.1
